structure

+++ fuckshadows-csharp/Encryption/Sodium.cs
@@ -63,19 +63,19 @@ namespace Fuckshadows.Encryption
             ulong inlen, byte[] key, int keylen, byte[] salt, byte[] personal);
 
         [DllImport(DLLNAME, CallingConvention = CallingConvention.Cdecl)]
-        public static extern int crypto_aead_chacha20poly1305_ietf_encrypt(byte[] c, ref ulong clen_p, byte[] m,
+        public static extern int crypto_aead_chacha20poly1305_ietf_encrypt(byte[] c, ref int clen_p, byte[] m,
             int mlen, IntPtr ad, int adlen, IntPtr nsec, byte[] npub, byte[] k);
 
         [DllImport(DLLNAME, CallingConvention = CallingConvention.Cdecl)]
-        public static extern int crypto_aead_chacha20poly1305_ietf_decrypt(byte[] m, ref ulong mlen_p,
+        public static extern int crypto_aead_chacha20poly1305_ietf_decrypt(byte[] m, ref int mlen_p,
             IntPtr nsec, byte[] c, int clen, IntPtr ad, int adlen, byte[] npub, byte[] k);
 
         [DllImport(DLLNAME, CallingConvention = CallingConvention.Cdecl)]
-        public static extern int crypto_aead_chacha20poly1305_encrypt(byte[] c, ref ulong clen_p, byte[] m, int mlen,
+        public static extern int crypto_aead_chacha20poly1305_encrypt(byte[] c, ref int clen_p, byte[] m, int mlen,
             IntPtr ad, int adlen, IntPtr nsec, byte[] npub, byte[] k);
 
         [DllImport(DLLNAME, CallingConvention = CallingConvention.Cdecl)]
-        public static extern int crypto_aead_chacha20poly1305_decrypt(byte[] m, ref ulong mlen_p, IntPtr nsec, byte[] c,
+        public static extern int crypto_aead_chacha20poly1305_decrypt(byte[] m, ref int mlen_p, IntPtr nsec, byte[] c,
             int clen, IntPtr ad, int adlen, byte[] npub, byte[] k);
 
         #endregion
